+++ packages/input/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 1.2.6
+
+### Patch Changes
+
+- Updated dependencies
+  [[`620f0b7d7`](https://github.com/chakra-ui/chakra-ui/commit/620f0b7d756ffb6bfc6ddf0459e96f774ffbb9be)]:
+  - @chakra-ui/form-control@1.3.6
+
 ## 1.2.5
 
 ### Patch Changes