+++ website/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 1.2.6
+
+### Patch Changes
+
+- Updated dependencies
+  [[`ddd5ef4a1`](https://github.com/chakra-ui/chakra-ui/commit/ddd5ef4a1e9cc988c99b80c26579205ea4c57b2f)]:
+  - @chakra-ui/react@1.6.2
+  - @chakra-ui/props-docs@1.0.27
+
 ## 1.2.5
 
 ### Patch Changes